# Balcony/src/main/resources/BalconySwagger.yaml
swagger: '2.0'
info:
  description: >-
    This is a sample server Balcony server.
  version: 1.0.0
  title: Swagger Balcony
host: petstore.swagger.io
basePath: /v2
...
schemes:
  - http
paths:  
  /data:
    get:
      tags:
        - BalconyController
      summary: Get all data
      description: Returns all data
      operationId: getAllData
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/SensorsResponseEntity'
        '400':
          description: Invalid ID supplied
        '404':
          description: Pet not found
  /waterPump:
    patch:
      tags:
        - BalconyController
      summary: Water pump control
      description: Enable/Dissable water pump.
      operationId: patchWaterPump
      produces:
        - application/json
      parameters:
        - name: enabled
          in: header
          required: true
          type: boolean  
        - name: duration
          in: header
          required: false
          type: integer    
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/BooleanReponseDefinition'
        '400':
          description: Invalid ID supplied
        '404':
          description: Pet not found         
...
definitions:
  SensorsResponseEntity:
    type: object
    properties:
      id:
        type: integer
        format: int64
      dataResponse:
        $ref: '#/definitions/DataResponse'    
  DataResponse:
    type: object
    properties:
      temperature:
        type: integer
        format: int64
      press:
        type: integer
        format: int64
      alt:
        type: integer
        format: int64
      soilStatus:
        type: boolean
      rainStatus:
        type: boolean
      waterPumpStatus:
        type: boolean
  BooleanReponseDefinition:
        type: object
        properties:
          enabled:
            type: boolean
...